Remove raw data processing from GetDataService and create abstract classes for querying raw observation and procedure data. The ProcedureOutputAdapter and TimeseriesOutputAdapter extends the new abstract classes.
