+++ sensorwebclient-sos-rest/src/main/java/org/n52/series/api/proxy/v1/srv/TimeseriesOutputAdapter.java
@@ -30,7 +30,6 @@ package org.n52.series.api.proxy.v1.srv;
 import static org.n52.series.api.proxy.v1.srv.QueryParameterAdapter.createQueryParameters;
 import static org.n52.server.mgmt.ConfigurationContext.getSOSMetadatas;
 
-import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
@@ -43,7 +42,6 @@ import org.n52.io.format.TvpDataCollection;
 import org.n52.io.v1.data.TimeseriesMetadataOutput;
 import org.n52.io.v1.data.UndesignedParameterSet;
 import org.n52.sensorweb.v1.spi.ParameterService;
-import org.n52.sensorweb.v1.spi.RawDataService;
 import org.n52.sensorweb.v1.spi.TimeseriesDataService;
 import org.n52.series.api.proxy.v1.io.TimeseriesConverter;
 import org.n52.shared.requests.query.QueryParameters;
@@ -51,7 +49,7 @@ import org.n52.shared.serializable.pojos.sos.SOSMetadata;
 import org.n52.shared.serializable.pojos.sos.SosTimeseries;
 import org.n52.shared.serializable.pojos.sos.Station;
 
-public class TimeseriesOutputAdapter implements TimeseriesDataService, ParameterService<TimeseriesMetadataOutput>, RawDataService {
+public class TimeseriesOutputAdapter extends RawObservationDataService implements TimeseriesDataService, ParameterService<TimeseriesMetadataOutput> {
 
     private GetDataService dataService;
 
@@ -169,25 +167,25 @@ public class TimeseriesOutputAdapter implements TimeseriesDataService, Parameter
 		this.statusIntervalsService = statusIntervalsService;
 	}
 
-	@Override
-	public InputStream getRawData(String id, IoParameters query) {
-		if (dataService instanceof RawDataService) {
-			return ((RawDataService)dataService).getRawData(id, query);
-		}
-		return null;
-	}
-
-	@Override
-	public InputStream getRawData(UndesignedParameterSet parameters) {
-		if (dataService instanceof RawDataService) {
-			return ((RawDataService)dataService).getRawData(parameters);
-		}
-		return null;
-	}
-
-	@Override
-	public boolean supportsRawData() {
-		return dataService instanceof RawDataService && ((RawDataService)dataService).supportsRawData();
-	}
+//	@Override
+//	public InputStream getRawData(String id, IoParameters query) {
+//		if (dataService instanceof RawDataService) {
+//			return ((RawDataService)dataService).getRawData(id, query);
+//		}
+//		return null;
+//	}
+//
+//	@Override
+//	public InputStream getRawData(UndesignedParameterSet parameters) {
+//		if (dataService instanceof RawDataService) {
+//			return ((RawDataService)dataService).getRawData(parameters);
+//		}
+//		return null;
+//	}
+//
+//	@Override
+//	public boolean supportsRawData() {
+//		return dataService instanceof RawDataService && ((RawDataService)dataService).supportsRawData();
+//	}
 
 }
